resa/backend/requirements.txt
google-generativeai>=0.5.0
openai>=1.14.0
python-dotenv>=1.0.1
PyPDF2>=3.0.1
fastapi
uvicorn
python-multipart
requests
python-pptx
tavily-python>=0.3.0
